Add feature to filter by extension option (--extensions)
New option --extensions can be used to filter by extension so that only files with the specified extension(s) will be searched in the source directory. ExifTool already includes the capability to include file extensions and this change utilizes that functionality.
Update readme to show the commented out feature for --extensions option.